# Added optimized vashishta potential.