As discussed today, an optimized vashishta implementation with 2 optimizations. Both optimizations increases performance by a total factor 3.

1) tabulated 2-body forces. Simulation time (in.sio2 with 10k timesteps) went from 155.8 secs to 80.51 secs. See attached .zip-file for output. 
2) building a new short range 3-body neighbor list when looping through 2-body neighborlist. Since 3-body cutoff is 2.6Å and 2-body is 10Å for SiO2, we can skip many 3-body triplets that are out of cutoff. Simulation time went from 80.51 secs to 51.71 secs.

I also added a pair_modify table yes/no [tableSize] command. USER-OMP version is also updated. 

[vashishta_benchmark.zip](https://github.com/lammps/lammps/files/421704/vashishta_benchmark.zip)