@andeplane 2016-08-17
The implementation obviously needs code review and testing, but I know it is an insult to even mention that, so I won't :p