@andeplane 2016-08-17
If using hybrid pair style, the pair_modify might corrupt other pair styles?