@akohlmey 2016-08-17
It would be nice to have a similar syntax in specifying the table size to what is used for coulomb, i.e. table size 0 means off, number of entries is 2^n (default 12, i.e. 2^12 = 4096). coulomb is single precision only though, due using floating point mantissa lookup tricks. see pair.cpp for details.