@andeplane 2016-08-22
It now uses 'pair_modify table N' where table size is 2^N. I also added the 'pair_modify tabinner r' command. We could probably add this float->int conversion trick you mentioned, but I haven't looked into it yet. It would be interesting to see if there are any notable performance increase.

Table is now disabled by default, and any non-zero N in 'pair_modify table N' enables the mode.