@andeplane 2016-08-22
I used lammps-icms branch in this request, and you cannot (afaik) change branch. So I opened a new one here:

https://github.com/lammps/lammps/pull/150