@akohlmey 2016-08-22
At the moment, all pull requests will be merged into lammps-icms. Using upstream as reference is not a big difference, but will cause be a couple more extra steps, e.g. manually closing the pull request, since integration into upstream means a new commit hash created by subversion and thus github cannot detect that a branch was merged with the target branch.